Rename message.py to handler.py
